ECKIT-620 Enable multi-value options in CmdArgs
In some ECFLOW cases, command line options are expected to collect multiple values, as if these values where positional relative to the option e.g. --option <command> <name> <value> /path/to/one /path/to/two ... --next. The logic mandates that values are collected until: 1) there are no more values to collect, either because the end of the list of values was reached or because the next option was found; or, 2) a mandatory+optional number of values has been collected.

This implies changes to the way Options are collected by CmdArgs, and the current change enables each kind of option (SimpleOption, VectorOption, etc) to define how and how many argv entries are collected.