ECKIT-620 Allow definition of default value for option
The ability to set an option's default value using a data member existed before, but was removed in favor of definition of the default value at construction.

Unfortunately, downstream dependencies are using this feature and so the feature was brought back, but only implemented for the classes for which it make sense (e.g. SimpleOption<t> provides it by inheriting it from BaseOption<T>; Separator does not provide).