ECKIT-631 eckit::Fraction improved convergence test fix
This fixes an issue with eckit::Fraction on some specific conversions from real values (double). It incrementally relaxes the comparison to determine fraction convergence, a more robust approach in the general case.

Note that the original behaviour is also preserved, which is just the first attempt to converge strictly (without relaxation), so this does not incur on any changes of results for already working code (particularly mir.) Instead it only makes some additional specific cases work (found on handling some unstructured grids.)