Remove use of memset for non trivially_copyable
Replaces the use of zeroing memory over all of TaskInfos memory from
within the constructor to value initialization of all members an
transitive members.

An initial fix attempt did break binary compatibility of TaskInfo types,
which are written to disk. Unfortuntely we did not have any test who
alerted us of this breakage.

This commit also introduces a new test that ensures that a TaskInfo type
initialized though placement new does initialialize all his members
properly. For this test to be possible it was required to refactor
the TaskInfo -> Padded<Info, 4096> relationship from inheritance to
composition. All fields of Info are now public and can be accessed for
testing. Info is fully contained in TaskInfo and the TaskInfo interface
is unchanged.